third_party/blink/web_tests/platform/mac-mac15-arm64/virtual/gpu/fast/canvas/canvas-imageSmoothingQuality-expected.txt

CONSOLE WARNING: Canvas2D: Multiple readback operations using getImageData are faster with the willReadFrequently attribute set to true. See: https://html.spec.whatwg.org/multipage/canvas.html#concept-canvas-will-read-frequently
This is a testharness.js-based test.
[FAIL] Tests for the imageSmoothingQuality attribute.
  assert_false: expected false got true
Harness: the test ran to completion.
